Clarify what timezone Slot timestamps should use
Clarify that the “local timezone” that Slot timestamps should use is the timezone of the Slot’s Location. The previous language implied this, but didn’t quite make it 100% clear that we were referring to the Slot’s Location and not, for example, the requester’s timezone or the managing organization’s office timezone.