Add section with guidelines for Slot Aggregators
We’ve had a lot of discussion in FHIR chat and on calls about unusual situations aggregators encounter, but that should not apply to first-party slot publishers. In order to provide clearer guidelines and aggregator-centric features but not encourage publishers to abuse them, I’ve added a separate section describing specific features and patterns for aggregators.